requirements.txt
python-dotenv
google-genai
gradio
pillow
gtts
pyttsx3
pypiwin32
SpeechRecognition
pydub
